REFACTOR: Move tracer utils to their own file.
