+++ eth/tracers/blocknative/utils.go
@@ -1,0 +1,27 @@
+package blocknative
+
+import (
+	"github.com/ethereum/go-ethereum/common"
+	"math/big"
+	"strconv"
+	"strings"
+)
+
+func bytesToHex(s []byte) string {
+	return "0x" + common.Bytes2Hex(s)
+}
+
+func bigToHex(n *big.Int) string {
+	if n == nil {
+		return ""
+	}
+	return "0x" + n.Text(16)
+}
+
+func uintToHex(n uint64) string {
+	return "0x" + strconv.FormatUint(n, 16)
+}
+
+func addrToHex(a common.Address) string {
+	return strings.ToLower(a.Hex())
+}
